requirements.txt
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache2==0.2.2
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
kombu==5.5.1
makefun==1.15.6
Mako==1.3.9
MarkupSafe==3.0.2
pendulum==3.0.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==4.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
time-machine==2.16.0
typing_extensions==4.13.0
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
